--- gg-calendar-api/src/main/java/calendar/gg/calendar/api/schedule/privateschedule/controller/request/PrivateScheduleReqDto.java
@@ -1,4 +1,0 @@
-package calendar.gg.calendar.api.schedule.privateschedule.controller.request;
-
-public class PrivateScheduleReqDto {
-}
