--- gg-calendar-api/src/main/java/calendar/gg/calendar/api/schedule/privateschedule/controller/response/PrivateScheduleResDto.java
@@ -1,4 +1,0 @@
-package calendar.gg.calendar.api.schedule.privateschedule.controller.response;
-
-public class PrivateScheduleResDto {
-}
