--- gg-calendar-api/src/main/java/calendar/gg/calendar/api/schedule/privateschedule/service/PrivateScheduleService.java
@@ -1,4 +1,0 @@
-package calendar.gg.calendar.api.schedule.privateschedule.service;
-
-public class PrivateScheduleService {
-}
