--- gg-calendar-api/src/main/java/calendar/gg/calendar/api/schedule/publicschedule/controller/request/PublicScheduleReqDto.java
@@ -1,4 +1,0 @@
-package calendar.gg.calendar.api.schedule.publicschedule.controller.request;
-
-public class PublicScheduleReqDto {
-}
