--- gg-calendar-api/src/main/java/calendar/gg/calendar/api/schedule/publicschedule/controller/response/PublicScheduleResDto.java
@@ -1,4 +1,0 @@
-package calendar.gg.calendar.api.schedule.publicschedule.controller.response;
-
-public class PublicScheduleResDto {
-}
