--- gg-calendar-api/src/main/java/calendar/gg/calendar/api/schedule/publicschedule/service/PublicScheduleService.java
@@ -1,4 +1,0 @@
-package calendar.gg.calendar.api.schedule.publicschedule.service;
-
-public class PublicScheduleService {
-}
