--- gg-calendar-api/src/main/java/calendar/gg/calendar/api/util/AuthVerification.java
@@ -1,4 +1,0 @@
-package calendar.gg.calendar.api.util;
-
-public class AuthVerification {
-}
