+++ gg-calendar-api/src/main/java/gg/calendar/api/schedule/privateschedule/controller/PrivateScheduleController.java
@@ -1,0 +1,4 @@
+package gg.calendar.api.schedule.privateschedule.controller;
+
+public class PrivateScheduleController {
+}
