+++ gg-calendar-api/src/main/java/gg/calendar/api/schedule/privateschedule/service/PrivateScheduleService.java
@@ -1,0 +1,4 @@
+package gg.calendar.api.schedule.privateschedule.service;
+
+public class PrivateScheduleService {
+}
