+++ gg-calendar-api/src/main/java/gg/calendar/api/schedule/publicschedule/controller/PublicScheduleController.java
@@ -1,0 +1,4 @@
+package gg.calendar.api.schedule.publicschedule.controller;
+
+public class PublicScheduleController {
+}
