+++ gg-calendar-api/src/main/java/gg/calendar/api/schedule/publicschedule/controller/request/PublicScheduleReqDto.java
@@ -1,0 +1,4 @@
+package gg.calendar.api.schedule.publicschedule.controller.request;
+
+public class PublicScheduleReqDto {
+}
