+++ gg-calendar-api/src/main/java/gg/calendar/api/schedule/publicschedule/controller/response/PublicScheduleResDto.java
@@ -1,0 +1,4 @@
+package gg.calendar.api.schedule.publicschedule.controller.response;
+
+public class PublicScheduleResDto {
+}
