+++ gg-calendar-api/src/main/java/gg/calendar/api/schedule/publicschedule/service/PublicScheduleService.java
@@ -1,0 +1,4 @@
+package gg.calendar.api.schedule.publicschedule.service;
+
+public class PublicScheduleService {
+}
