+++ gg-calendar-api/src/main/java/gg/calendar/api/util/AuthVerification.java
@@ -1,0 +1,4 @@
+package gg.calendar.api.util;
+
+public class AuthVerification {
+}
